refactor(payment): FeignConfig 리팩터링 및 모듈 구조 변경
- Feign 설정 클래스 위치 변경 (`modules/feign` → `commerce-api`)
- `modules:feign` 모듈 삭제 및 관련 의존성 `commerce-api`로 이동
- `PgException` 클래스 정리 및 이름 변경 (`PgInfraException` → 구체적 예외)
- 코드 단순화 및 비즈니스 로직에서 예외 처리 전략 개선

모듈 구조 간소화 및 Feign 설정 위치 명확화
