+++ apps/commerce-api/src/main/kotlin/com/loopers/infrastructure/payment/PgFeignClient.kt
@@ -1,7 +1,6 @@
 package com.loopers.infrastructure.payment
 
 import org.springframework.cloud.openfeign.FeignClient
-import org.springframework.stereotype.Component
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestParam
     url = "\${pg.base-url}",
     configuration = [PgFeignConfig::class],
 )
-@Component
 interface PgFeignClient {
 
     @PostMapping("/api/v1/payments")
