--- modules/feign/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    api("org.springframework.cloud:spring-cloud-starter-openfeign")
-}
