+++ settings.gradle.kts
@@ -7,7 +7,6 @@ include(
     ":modules:jpa",
     ":modules:redis",
     ":modules:kafka",
-    ":modules:feign",
     ":supports:jackson",
     ":supports:logging",
     ":supports:monitoring",
